TripoSR/requirements.txt
omegaconf==2.3.0
opencv-python-headless>=4.8.0
open3d>=0.18.0
Pillow==10.1.0
einops==0.7.0
git+https://github.com/tatsy/torchmcubes.git
transformers==4.35.0
trimesh==4.0.5
rembg
huggingface-hub
imageio[ffmpeg]
gradio
xatlas==0.0.9
moderngl==5.10.0
